backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pymongo
pydantic
numpy
faiss-cpu